requirements.txt
# Prod
PySide6>=6.2.4
psutil>=5.9.1
pillow>=9.1.1
proxy.py>=2.4.2
peewee>=3.14.10
plyer>=2.0.0
pygame>=2.1.2
grabicon>=0.1.2

# Dev
flake8>=4.0.1
pytest>=7.1.1
pyinstaller>=5.1
Nuitka>=0.8.1
pycodestyle<2.9.0,>=2.8.0
black>=23.3.0 